added people properties
