+++ app/profiles/services/update_team.rb
@@ -40,7 +40,7 @@ module Services
 
     def update_to_mixpanel
       if succeeded?
-        mixpanel = MixpanelTracker.new user: team.user
+        mixpanel = Services::MixpanelTracker.new user: team.user
         mixpanel.add_team
       end
     end
